Add CRUD operations for Contact and Review entities
- Implemented create, store, show, edit, update, and destroy methods for Contact and Review controllers, enhancing the backend functionality for managing these entities.
- Added validation rules for storing and updating Contact and Review data to ensure data integrity.
- Introduced views for creating and editing Contacts and Reviews, improving the user interface for backend management.

+++ app/Http/Controllers/Backend/Admin/OrderController.php
@@ -31,6 +31,7 @@ class OrderController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('order_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $orders = Order::get();
 
         return view('backend.orders.index', compact('orders'));
@@ -480,6 +481,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function create()
     {
+        abort_if(Gate::denies('order_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $users = User::orderBy('first_name')->get();
         
         $courses = Course::where('published', 1)->orderBy('title')->get();
@@ -495,6 +497,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('order_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $request->validate([
             'user_id' => 'required|exists:users,id',
             'course_ids' => 'required|array|min:1',
@@ -617,6 +620,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function edit($id)
     {
+        abort_if(Gate::denies('order_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $order = Order::findOrFail($id);
         $users = User::orderBy('first_name')->get();
         
@@ -637,6 +641,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function update(Request $request, $id)
     {
+        abort_if(Gate::denies('order_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $order = Order::findOrFail($id);
         
         $request->validate([
@@ -760,6 +765,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function show($id)
     {
+        abort_if(Gate::denies('order_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $order = Order::findOrFail($id);
         // dd($order);
         //  generateInvoice($order);
@@ -779,7 +785,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function destroy($id)
     {
-
+        abort_if(Gate::denies('order_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $order = Order::findOrFail($id);
         $order->items()->delete();
         $order->delete();
@@ -793,9 +799,7 @@ $orders->where("end_date","<=",date("Y-m-d"));
      */
     public function massDestroy(Request $request)
     {
-        if (!Gate::allows('course_delete')) {
-            return abort(401);
-        }
+        abort_if(Gate::denies('order_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->input('ids')) {
             $entries = Order::whereIn('id', $request->input('ids'))->get();
             foreach ($entries as $entry) {
